single responsibility principle - callChatFunctions.ts
