+++ evals/first-week-album-sales.eval.ts
@@ -1,10 +1,6 @@
 import { Eval } from "braintrust";
 import { Factuality } from "autoevals";
-import { UIMessage } from "ai";
-import { DEFAULT_MODEL } from "@/lib/consts";
-import { setupChatRequest } from "@/lib/chat/setupChatRequest";
-import { generateText } from "ai";
-import { type ChatRequest } from "@/lib/chat/types";
+import { callChatFunctions } from "@/lib/evals";
 
 /**
  * Album Sales Tool Call Evaluation
@@ -12,18 +8,10 @@ import { type ChatRequest } from "@/lib/chat/types";
  * This evaluation tests whether your AI system properly uses tool calls
  * to fetch album sales data instead of defaulting to "I don't have access" responses.
  *
- * This evaluation calls the chat functions directly (setupChatRequest + generateText)
- * instead of making HTTP requests, making it faster and more reliable.
- *
- * Run: npx braintrust eval evals/album-sales-tool-call.eval.ts
+ * Run: npx braintrust eval evals/first-week-album-sales.eval.ts
  */
 
-// Note: The Factuality scorer from autoevals uses GPT-4o by default
-// To use GPT-5-mini, set the OPENAI_MODEL environment variable:
-// export OPENAI_MODEL=gpt-5-mini
-// Then run: npx braintrust eval evals/album-sales-tool-call.eval.ts
-
-Eval("Album Sales Tool Call Evaluation", {
+Eval("First Week Album Sales Evaluation", {
   data: () => [
     {
       input:
@@ -34,7 +22,7 @@ Eval("Album Sales Tool Call Evaluation", {
         artist: "Halsey",
         album: "The Great Impersonator",
         expected_tool_usage: true,
-        data_type: "first_week_sales",
+        data_type: "first_week_album_sales",
       },
     },
     {
@@ -46,7 +34,7 @@ Eval("Album Sales Tool Call Evaluation", {
         artist: "Taylor Swift",
         album: "Midnights",
         expected_tool_usage: true,
-        data_type: "first_week_sales",
+        data_type: "first_week_album_sales",
       },
     },
     {
@@ -58,7 +46,7 @@ Eval("Album Sales Tool Call Evaluation", {
         artist: "Drake",
         album: "Certified Lover Boy",
         expected_tool_usage: true,
-        data_type: "first_week_sales",
+        data_type: "first_week_album_sales",
       },
     },
     {
@@ -70,7 +58,7 @@ Eval("Album Sales Tool Call Evaluation", {
         artist: "Bad Bunny",
         album: "Un Verano Sin Ti",
         expected_tool_usage: true,
-        data_type: "streaming_numbers",
+        data_type: "first_week_streaming_numbers",
       },
     },
   ],
@@ -87,52 +75,3 @@ Eval("Album Sales Tool Call Evaluation", {
 
   scores: [Factuality],
 });
-
-// Call the chat functions directly instead of making HTTP requests
-async function callChatFunctions(input: string): Promise<string> {
-  // Create the message structure expected by your chat system
-  const messages: UIMessage[] = [
-    {
-      id: "user-message",
-      role: "user",
-      parts: [
-        {
-          type: "text",
-          text: input,
-        },
-      ],
-    },
-  ];
-
-  // Prepare the request body matching your ChatRequest type
-  const body: ChatRequest = {
-    messages,
-    roomId: "3779c62e-7583-40c6-a0bb-6bbac841a531",
-    accountId: "fb678396-a68f-4294-ae50-b8cacf9ce77b",
-    model: DEFAULT_MODEL,
-    excludeTools: [], // Don't exclude any tools - we want to test tool usage
-  };
-
-  try {
-    // Use the same functions as your API endpoint
-    const chatConfig = await setupChatRequest(body);
-    const result = await generateText(chatConfig);
-
-    // Return the text content from the result
-    // Handle different possible response formats
-    if (typeof result.text === "string") {
-      return result.text;
-    }
-
-    // Try to extract text from content if available
-    if (typeof result.content === "string") {
-      return result.content;
-    }
-
-    // Fallback - convert result to string
-    return String(result.text || result.content || "No response content");
-  } catch (error) {
-    console.error("Error calling chat functions:", error);
-    throw error;
-  }
-}
